cs_insn.address also needs to be changed after the change on @offset of cs_disasm*() API

+++ include/capstone.h
@@ -52,7 +52,7 @@ typedef struct cs_insn {
 	unsigned int id;
 
 	// Offset address of this instruction
-	size_t address;
+	uint64_t address;
 
 	// Size of this instruction
 	uint16_t size;
